+++ gammapy/datasets/core.py
@@ -77,6 +77,20 @@ class Dataset(abc.ABC):
 
         return np.sum(stat, dtype=np.float64)
 
+    def set_priors_from_models(self, models):
+        if models is not None:
+            self._priors = [
+                par.prior for par in models.parameters if par.prior is not None
+            ]
+        else:
+            self._priors = None
+
+    @property
+    def priors(self):
+        """Priors set on model parameters (list)."""
+        self.set_priors_from_models(self.models)
+        return self._priors
+
     @abc.abstractmethod
     def stat_array(self):
         """Statistic array, one value per data point."""
