+++ gammapy/datasets/map.py
@@ -358,20 +358,6 @@ class MapDataset(Dataset):
                     self._evaluators[model.name] = evaluator
         self._models = models
 
-    def set_priors_from_models(self, models):
-        if models is not None:
-            self._priors = [
-                par.prior for par in models.parameters if par.prior is not None
-            ]
-        else:
-            self._priors = None
-
-    @property
-    def priors(self):
-        """Priors set on model parameters (list)."""
-        self.set_priors_from_models(self.models)
-        return self._priors
-
     @property
     def evaluators(self):
         """Model evaluators."""
